neural_swarm/requirements.txt
fastapi
uvicorn
google-genai
pydantic
python-multipart
wave
dataclasses
